rename file, remove unnecessary protocols
